+++ tests/assets/sub-path-script-error/js/script.js
@@ -1,0 +1,9 @@
+function run () {
+  fetch('http://test/json/foo')
+    .catch(e => {
+      console.error(e)
+      newrelic.noticeError(e)
+    })
+}
+
+run()