+++ tools/testing-server/routes/bam-apis.js
@@ -6,7 +6,15 @@ const { rumFlags } = require('../constants')
  * @param {module:fastify.FastifyInstance} fastify the fastify server instance
  * @param {TestServer} testServer test server instance
  */
-module.exports = fp(async function (fastify, testServer) {
+module.exports = fp(async function (fastify) {
+  fastify.route({
+    method: ['GET', 'POST'],
+    url: '/debug',
+    handler: async function (request, reply) {
+      fastify.testServerLogger.logDebugShimMessage(request)
+      reply.code(200).send()
+    }
+  })
   fastify.route({
     method: ['GET', 'POST'],
     url: '/1/:testId',