+++ tools/wdio/plugins/jil-commands.mjs
@@ -1,7 +1,7 @@
 /**
  * This is a WDIO worker plugin that provides custom commands.
  */
-export default class TestingServerWorker {
+export default class JilCommands {
   async before (capabilities, context, browser) {
     browser.addCommand('waitForFeature', async function (feature) {
       const command = `window.NREUM && window.NREUM.activatedFeatures && window.NREUM.activatedFeatures['${feature}']`
@@ -11,7 +11,7 @@ export default class TestingServerWorker {
           return window.NREUM && window.NREUM.activatedFeatures && window.NREUM.activatedFeatures[feat]
         }, feature),
         {
-          timeout: 60000,
+          timeout: 30000,
           timeoutMsg: 'Agent never loaded'
         })
     })